--- UploadServlet.java~
@@ -1,43 +1,0 @@
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.MultipartConfig;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.Part;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.BufferedWriter;
-import java.io.PrintWriter;
-
-@WebServlet("/UploadServlet")
-@MultipartConfig(location="/tmp", maxFileSize=1048576)
-public class UploadServlet extends HttpServlet {
-  @Override
-  protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    Part part = request.getPart("file");
-    File file = new File("check.txt");
-    if(checkBeforeWritefile(file)){
-      PrintWriter pw = new PrintWriter(new BufferedWriter(new FileWriter(file)));
-      pw.println(part);
-
-      pw.close();
-    }
-    String name = this.getFileName(part);
-    part.write(getServletContext().getRealPath("/usr/local/MyApp/webpro/WebPro2016_6/image") + "/" + name);
-    response.sendRedirect("/usr/local/MyApp/circle_regist.jsp");
-  }
-
-  private String getFileName(Part part) {
-    String name = null;
-    for (String dispotion : part.getHeader("Content-Disposition").split(";")) {
-      if (dispotion.trim().startsWith("filename")) {
-        name = dispotion.substring(dispotion.indexOf("=") + 1).replace("\"", "").trim();
-        name = name.substring(name.lastIndexOf("\\") + 1);
-        break;
-      }
-    }
-    return name;
-  }
-}
