+++ WEB-INF/web.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<web-app xmlns="http://java.sun.com/xml/ns/j2ee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee
+         http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd"
+         version="2.4">
+
+<servlet>
+  <servlet-name>ImgUpSv</servlet-name>
+  <servlet-class>utility.ImageUploader</servlet-class>
+</servlet>
+
+<servlet-mapping>
+    <servlet-name>ImgUpSv</servlet-name>
+    <url-pattern>/imguploader</url-pattern>
+</servlet-mapping>
+
+
+
+
+</web-app>
