--- top_test.html
@@ -1,17 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<title>トップページ(仮)</title>
-</head>
-<body>
-<form action=circle_regist.jsp>
-<input type="submit" value="サークル登録">
-</form>
-<form action=/*サークル検索*/>
-<input type="submit" value="サークル検索">
-</form>
-<form action=circle_admin.jsp>
-<input type="submit" value="サークル管理">
-</form>
-</body>
-</html>
